refactor: rename InputView folder to StakeInputView
