+++ app/components/UI/Stake/Views/StakeInputView/StakeInputView.styles.ts
