+++ app/components/UI/Stake/Views/StakeInputView/StakeInputView.test.tsx
