+++ app/components/UI/Stake/Views/StakeInputView/StakeInputView.tsx
