+++ app/components/UI/Stake/routes/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import StakeInputView from '../Views/InputView/StakeInputView';
+import StakeInputView from '../Views/StakeInputView/StakeInputView';
 import LearnMoreModal from '../components/LearnMoreModal';
 import Routes from '../../../../constants/navigation/Routes';
 import UnstakeInputView from '../Views/UnstakeInputView/UnstakeInputView';
